+++ sdk/core/azure_core/src/lib.rs
@@ -36,26 +36,6 @@ pub mod tracing {
 #[cfg(feature = "xml")]
 pub use typespec_client_core::xml;
 
-#[cfg(not(target_arch = "wasm32"))]
-mod conditional_send {
-    /// Conditionally implements [`Send`] based on the `target_arch`.
-    ///
-    /// This implementation requires `Send`.
-    pub trait ConditionalSend: Send {}
-
-    impl<T> ConditionalSend for T where T: Send {}
-}
-
-#[cfg(target_arch = "wasm32")]
-mod conditional_send {
-    /// Conditionally implements [`Send`] based on the `target_arch`.
-    ///
-    /// This implementation does not require `Send`.
-    pub trait ConditionalSend {}
-
-    impl<T> ConditionalSend for T {}
-}
-
 mod private {
     pub trait Sealed {}
 }