+++ sdk/eventhubs/azure_messaging_eventhubs/src/common/recoverable/sender.rs
@@ -42,8 +42,7 @@ impl RecoverableSender {
     }
 }
 
-#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+#[async_trait::async_trait]
 impl AmqpSenderApis for RecoverableSender {
     async fn send<M>(&self, message: M, options: Option<AmqpSendOptions>) -> Result<AmqpSendOutcome>
     where